Show email in header again, except on mobile
Refactor the sign in modal to its own file
